Migrate from SQLite to PostgreSQL
- Replace sqlx sqlite feature with postgres
- Convert all ? bind params to $1, $2... positional params
- Replace datetime('now') with formatted chrono strings
- Convert AUTOINCREMENT to SERIAL, use BIGINT for github_id/cycles
- Keep TEXT columns for timestamps (compatible with String model fields)
- Add databases section to render.yaml for managed Postgres
- Remove ephemeral SQLite data dir from Dockerfile
- Add SQLX_OFFLINE=true to Dockerfile build stage
- Delete duplicate migration file

+++ backend/src/models.rs
@@ -59,7 +59,7 @@ pub struct DeploymentRecord {
 
 #[derive(Debug, Serialize, Deserialize, sqlx::FromRow)]
 pub struct DeployLog {
-    pub id: i64,
+    pub id: i32,
     pub deployment_id: String,
     pub level: String,
     pub message: String,
